third

+++ src/main/java/com/example/test/service/TestService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.example.test.dao.TestDao;
 public interface TestService {
-   User findUserById(Long id);
+   User findUserById(Integer id);
